@@ -38,7 +38,7 @@
 		"nanostores": "^0.11.3",
 		"react": "^18",
 		"react-dom": "^18",
-		"svelte": "^4",
+		"svelte": "^5",
 		"tailwindcss": "^3.4.13",
 		"tailwindcss-spring": "^1.0.1",
 		"typescript": "^5.6.3",
