@@ -2,29 +2,45 @@
 	import NumberFlow, { type Format } from '@number-flow/svelte'
 	import clsx from 'clsx/lite'
 	import { Bookmark, ChartNoAxesColumn, Heart, Repeat, Share } from 'lucide-svelte'
+	import type { HTMLAttributes } from 'svelte/elements'
 
-	let cls = ''
-	export { cls as class }
-
-	export let likes: number
-	export let reposts: number
-	export let views: number
-	export let bookmarks: number
-	export let liked: boolean
-	export let reposted: boolean
-	export let bookmarked: boolean
-	export let onlike: () => void
-	export let onrepost: () => void
-	export let onbookmark: () => void
+	type Props = HTMLAttributes<HTMLDivElement> & {
+		likes: number
+		reposts: number
+		views: number
+		bookmarks: number
+		liked: boolean
+		reposted: boolean
+		bookmarked: boolean
+		onlike: () => void
+		onrepost: () => void
+		onbookmark: () => void
+	}
 
 	const format: Format = {
 		notation: 'compact',
 		compactDisplay: 'short',
 		roundingMode: 'trunc'
 	}
+
+	const {
+		likes,
+		reposts,
+		views,
+		bookmarks,
+		liked,
+		reposted,
+		bookmarked,
+		onlike,
+		onrepost,
+		onbookmark,
+		class: cls,
+		...props
+	}: Props = $props()
 </script>
 
 <div
+	{...props}
 	class={clsx(
 		cls,
 		'~text-[0.8125rem]/sm flex w-full select-none items-center text-zinc-600 dark:text-zinc-300'
@@ -38,7 +54,7 @@
 		<button
 			class="group flex items-center gap-1.5 pr-1.5 transition-[color] hover:text-emerald-500"
 			class:text-emerald-500={reposted}
-			on:click={onrepost}
+			onclick={onrepost}
 		>
 			<div
 				class="relative before:absolute before:-inset-2.5 before:rounded-full before:transition-[background-color] before:group-hover:bg-emerald-500/10"
@@ -55,7 +71,7 @@
 		<button
 			class="group flex items-center gap-1.5 pr-1.5 transition-[color] hover:text-pink-500"
 			class:text-pink-500={liked}
-			on:click={onlike}
+			onclick={onlike}
 		>
 			<div
 				class="relative before:absolute before:-inset-2.5 before:rounded-full before:transition-[background-color] before:group-hover:bg-pink-500/10"
@@ -75,7 +91,7 @@
 		<button
 			class="group flex items-center gap-1.5 pr-1.5 transition-[color] hover:text-blue-500"
 			class:text-blue-500={bookmarked}
-			on:click={onbookmark}
+			onclick={onbookmark}
 		>
 			<div
 				class="relative before:absolute before:-inset-2.5 before:rounded-full before:transition-[background-color] before:group-hover:bg-blue-500/10"
