@@ -1,21 +1,23 @@
 <script lang="ts">
-	import NumberFlow, { NumberFlowElement } from '@number-flow/svelte'
+	import NumberFlow from '@number-flow/svelte'
 	import clsx from 'clsx/lite'
 	import { Minus, Plus } from 'lucide-svelte'
 
-	export let min = 0
-	export let value = 0
-	export let max = 99
+	let {
+		min = 0,
+		value = $bindable(0),
+		max = 99
+	}: { min?: number; value?: number; max?: number } = $props()
 	const defaultValue = value
 
-	let flow: NumberFlowElement
 	let input: HTMLInputElement
 
+	let animated = $state(true)
 	// Hide the caret during transitions so you can't see it shifting around:
-	let showCaret = true
+	let showCaret = $state(true)
 
 	function handleInput() {
-		flow.animated = false // addresses race condition with updating
+		animated = false
 		let next = value
 		if (input.value === '') {
 			next = defaultValue
@@ -29,7 +31,7 @@
 	}
 
 	function handlePointerDown(event: PointerEvent, diff: number) {
-		flow.animated = true
+		animated = true
 		if (event.pointerType === 'mouse') {
 			event?.preventDefault()
 			input.focus()
@@ -47,7 +49,7 @@
 		tabindex={-1}
 		class="flex items-center pl-[.5em] pr-[.325em]"
 		disabled={min != null && value <= min}
-		on:pointerdown={(event) => handlePointerDown(event, -1)}
+		onpointerdown={(event) => handlePointerDown(event, -1)}
 	>
 		<Minus class="size-4" absoluteStrokeWidth strokeWidth="3.5" />
 	</button>
@@ -68,13 +70,13 @@
 			inputmode="numeric"
 			{max}
 			bind:value
-			on:input={handleInput}
+			oninput={handleInput}
 		/>
 		<NumberFlow
-			bind:el={flow}
 			{value}
 			format={{ useGrouping: false }}
 			aria-hidden
+			{animated}
 			on:animationsstart={() => (showCaret = false)}
 			on:animationsfinish={() => (showCaret = true)}
 			class="pointer-events-none"
@@ -86,7 +88,7 @@
 		tabindex="-1"
 		class="flex items-center pl-[.325em] pr-[.5em]"
 		disabled={max != null && value >= max}
-		on:pointerdown={(event) => handlePointerDown(event, 1)}
+		onpointerdown={(event) => handlePointerDown(event, 1)}
 	>
 		<Plus class="size-4" absoluteStrokeWidth strokeWidth="3.5" />
 	</button>
